fix_recommender/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-adk>=1.7.0
httpx>=0.24.0
redis>=4.5.0
google-generativeai==0.3.2
openai==1.6.1
pydantic==2.5.0
python-dotenv>=1.0.0
numpy>=1.24.0 